can now save token to the db

+++ src/main/kotlin/com/vauldex/inventory_management/controller/ProductController.kt
@@ -5,6 +5,7 @@ import com.vauldex.inventory_management.domain.dto.request.ProductRequest
 import com.vauldex.inventory_management.domain.dto.response.ProductResponse
 import com.vauldex.inventory_management.domain.entity.ProductEntity
 import com.vauldex.inventory_management.response.ResponseSuccess
+import com.vauldex.inventory_management.service.abstraction.AuthenticationService
 import com.vauldex.inventory_management.service.abstraction.ProductService
 import com.vauldex.inventory_management.utility.JwtUtils
 import org.springframework.http.HttpStatus
@@ -24,6 +25,7 @@ import java.util.UUID
 @RequestMapping("/api/products")
 class ProductController(
         private val prodService: ProductService,
+        private val authService: AuthenticationService,
         private val jwtUtils: JwtUtils
 ) {
 
@@ -35,7 +37,7 @@ class ProductController(
 
         val auth = authHeader.removePrefix("Bearer ").trim()
 
-        if(!jwtUtils.validateAccessToken(auth)) throw IllegalArgumentException("Invalid token.")
+        authService.validateAccessToken(auth)
 
         val category = prodService.getCategoryName(category = product.category)
         val prod = prodService.save(product = product.toEntity(cat = category))
@@ -48,7 +50,15 @@ class ProductController(
     }
 
     @GetMapping("/{productName}")
-    fun search(@PathVariable productName: String): ResponseSuccess<ProductResponse>{
+    fun search(
+            @PathVariable productName: String,
+            @RequestHeader("Authorization") authHeader: String
+    ): ResponseSuccess<ProductResponse>{
+
+        val auth = authHeader.removePrefix("Bearer ").trim()
+
+        authService.validateAccessToken(auth)
+
         val prod = prodService.search(product = productName)
 
         return ResponseSuccess(
@@ -59,7 +69,15 @@ class ProductController(
     }
 
     @PutMapping
-    fun edit(@RequestBody product: ProductEditRequest): ResponseSuccess<String> {
+    fun edit(
+            @RequestBody product: ProductEditRequest,
+            @RequestHeader("Authorization") authHeader: String
+    ): ResponseSuccess<String> {
+
+        val auth = authHeader.removePrefix("Bearer ").trim()
+
+        authService.validateAccessToken(auth)
+
         val category = prodService.getCategoryName(category = product.category)
         val prod = prodService.editProduct(product = product.toEntity(cat = category))
 
@@ -71,7 +89,15 @@ class ProductController(
     }
 
     @DeleteMapping
-    fun delete(@RequestParam idProduct: UUID): ResponseSuccess<String> {
+    fun delete(
+            @RequestParam idProduct: UUID,
+            @RequestHeader("Authorization") authHeader: String
+    ): ResponseSuccess<String> {
+
+        val auth = authHeader.removePrefix("Bearer ").trim()
+
+        authService.validateAccessToken(auth)
+
         val prod = prodService.deleteProduct(idProduct)
         val response = ResponseSuccess(
                 code = "PRODUCT_DELETED",
@@ -82,7 +108,11 @@ class ProductController(
     }
 
     @GetMapping
-    fun get(): ResponseSuccess<List<ProductEntity>> {
+    fun get(@RequestHeader("Authorization") authHeader: String): ResponseSuccess<List<ProductEntity>> {
+        val auth = authHeader.removePrefix("Bearer ").trim()
+
+        authService.validateAccessToken(auth)
+
         val productResponse = prodService.getAllProduct()
 
         return ResponseSuccess(
